"multiplication commit 002 - variable type changed int to double"

+++ src/main/java/com/softesteam/repodemo/calcoperations/Multiplication.java
@@ -4,7 +4,7 @@ public class Multiplication {
     public Multiplication() {
     }
 
-    public int multiplication(int a, int b) {
+    public double multiplication(double a, double b) {
         return a * b;
     }
 }
